backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-jose[cryptography]
pydantic
prometheus-api-client==0.6.0  # optional: helpful Prometheus client
python-dotenv
prometheus-fastapi-instrumentator
